Fix cacheing bug and comment.
The previous cacheing logic would not have considered the fact that an
entity's owner could be updated _to_ the current account's ID. So, if
access was not allowed, but then the owner was set to the account, it
would have no effect without a cache clear.

I removed the bit about the publishing status because that is not
relevant to that particular line. I did an the word "also" to indicate
that it operates in conjunction with the status logic above.